+++ CHANGELOG.md
@@ -1,9 +1,12 @@
 # CHANGELOG
 
 ## v0.1.2
-- Add labels for tutorials
+- Added labels for tutorials
     - in "TitleScreen" added "Move with A - D" label
     - in "Game" added "Press SPACE to jump" and "Beware the spikes!" labels
+- The "VersionLabel" in "TitleScreen" corresponds to "application/config/version" of the config
+- Added signal waves
+- The signal will reach if "player.can_interact == true"
 
 ## v0.1.1
 - Fixed the label "Press Q to propagate the signal" to "Hold Q to propagate the signal"